backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
pydantic-settings==2.3.4

# AI / RAG
google-generativeai==0.7.2
chromadb==0.5.5
sentence-transformers==3.0.1

# PDF processing and generation
pypdf==4.3.1
reportlab==4.2.2

# Utilities
python-multipart==0.0.9
httpx==0.27.0
